Add Extension Support
This patch adds basic extensions support. This was a pretty difficult change as extensions apply to the code generation request itself. Thus, we parse a first time to see all of the modules and extensions, we generate and run that code to load those modules, then we run the code generation request again. As an example, let's say that, for instance:

HttpRule extends MethodOptions

When we first parse the request, we need to track this as an extension (we do this in :ets). Secondly, we can't populate the HttpRule type yet because we don't have a module or definition for it. So we run through the code request on a first pass, track the extensions. Then we run the generated code (which defines the HttpRule module with it's __message_props__). Then we run the code request again with "run_extensions = true" which actually populates the HttpRule field appropriately. This allows us to then generate .ex files based on the fields of extensions.

This system allows us to properly add GrpcGateway information to the service module. Additionally, I refactored the service module to be agnostic to any particular service. It should be pretty simple for a library like Grpc to accept these simply modules and generate whatever code it desires.