Add descriptors to generated protos
There is a great deal of information in the descriptors that unfortunately is lost with the current approach.

Most protubuf libraries provide access to the raw descriptors, this PR brings this lib inline with that philosophy by
adding the encoded descriptor to the generated modules and exposing them via a `descriptor` function.