New feature: JSON decoding (#109)
This allows us to decode messages from JSON in addition to protobuf binaries.

It follows Google's specs and reference implementation. Only `proto3` syntax
is supported at the moment and some features, such as well-known types, are
not fully supported yet.

JSON decoding relies on the `Jason` library. Support to other JSON engines
may be added in the future.

https://developers.google.com/protocol-buffers/docs/proto3#json
https://developers.google.com/protocol-buffers/docs/reference/google.protobuf