Update Protobuf.Decoder: Fix decoding packed varints
Previously when decoding varints, the value was not correctly decoded
from wire format to protobuf format. Consequently when signed ints which
are encoded on the wire using ZigZag format, the values returned are not
correct. This commit updates the decoder to ensure all values are
transformed from wire format to protobuf format, and includes a couple
of unit tests to validate the expected behaviour.