Don't log warning when decoding unknown enums (#238)
Right now, the library warns if we decode an enum for which we don't know the **atom value**. That is, if you have an enum like this:

```protobuf
enum Status {
  LOGGED_IN = 0;
  LOGGED_OUT = 1;
}
```

and you decode it with a version of the schemas that don't have the `LOGGED_OUT = 1` definition, then the library is going to return `1` as the enum value. This is the correct thing to do to ensure backwards and forwards compatibility.

However, warning for this case is problematic. In many cases, applications want to handle these unknown enums *by design* and do their own thing or ignore the warning. Additionally, this is a 100% functional library so adding a "stateful" component (the logging) feels inappropriate.

When decoding enums, your business logic should always plan for unknown enums anyways, for backwards and forwards compatibility.

```elixir
case my_message.some_enum do
  :SOME_VALUE -> ...
  :SOME_OTHER_VALUE -> ...
  unknown -> ...
end
```

Since you need to handle this anyways, you might as well warn yourself if you are not equipped to handle unknown enums.