Adds .proto -> .pb.ex mix compiler
This library provides a `protoc` plugin, `protoc-gen-elixir`, for generating
elixir module out of protobuf definitions.

Due to API changes in `protobuf`, files generated with a previous version of
the plugin may not be compatible with this version of `protobuf`. This
compiler ensures elixir modules are generated with the same version of the
library.

In addition to that, the compiler provides more mix-friendly options:
* source directories are by default inferred from project configuration,
* destination directory is by default inferred from project configuration
* mandatory includes are enabled by default
* support for staled targets, proper clean, etc