Update Decoder: handle MatchError when decoding (#247)
Previously when decoding a delimited field, if the given binary has insufficient
data a MatchError was raised when attempting to pattern match on the
next value in the field. This commit updates the code to check the
length before attempting the match, and throws a Protobuf.DecodeError
with a more informative message.