Fix Dialyzer Errors
In order to pass `mix dialyzer` this commit defines the typespec for
__message_props__/0 within the body of the quote block, along with any
dependent types. Any attempt to use the remote types defined in the
other modules results in a `does not match the success typing` error.
It's not clear why this is happening, but maybe related to
the fact we're inside a "before_compile" macro.

The following typespec changes were also required to achieve a pass:

- MessageProps.extension_range could be nil
- FieldProps.type could be a tuple