Fix map fields contained in .proto files without a package (#325)
Closes #321.

In "nested_maps/2", the fully qualified name was built without discarding the
nil entries of the list. This meant that map fields worked correctly for .proto
files that contained a package, but failed to do so in .proto files that did
not, since the map key had then an extra "." in front that made the lookup in
"get_field/4" fail.