Make wire type public
This type is referenced from the typespecs generated by the protoc
extension. When building documentation for the protobufs (that is, using
include_docs=true), it can't be resolved and ex_doc will print the
following warning for every struct:

warning: documentation references type "Protobuf.Wire.Types.wire_type()" but
it is hidden or private.

By making it public, the warning is resolved, and it becomes clickable
on hexdocs.

Another alternative to solve the issue (while keeping the internal type
private) would be to change the typespec of `:__unknown_fields__` to avoid
referring to it, instead referring to a public type.