Add native Elixir `JSON` support (#394)
* Add `JSON` support

This implementation defaults to `JSON` if it's available. If it's not,
it tries to use `Jason` as it did before.

Some exceptions are changed (only for `JSON`, `Jason` is unaffected). For
example, instead of `Jason.DecodeError`, when using `JSON` we raise
`Protobuf.JSON.DecodeError`, and encode errors have a more varied shape
(since they can be mostly anything coming from Elixir side).

* Add Elixir 1.18.1 to workflow

* Add note

* fix typo

* Fix dialyzer

* Update lib/protobuf/json.ex