Add Protobuf.Text
* Allows encoding to textproto

* Run conformance for text, fixes for conformance, add unknown fields support

Most of the conformance tests in the text suite are for decoding, and
some of them are for unknown fields encoding rules, so we only got value
out of five of them. Still, they led to significant improvements in the
implementation.

For one, I figured that the "outer message" **can't** use brackets ("{",
"}"). I thought they were optional and just not used in the examples. So
we went from:

```elixir
Protobuf.Text.encode(%MyMessage{a: 1, b: 2})
```

To:
```elixir
Protobuf.Text.encode(%MyMessage{a: 1, b: 2})
```

I also added unknown fields support, but wasn't able to pass all the
unknown fields tests, so I skipped them for now. The current version
does the job of printing them reasonably well, so I'm not too bothered
with it.